rolevate-interview/requirements.txt
# Rolevate Interview Agent Dependencies - v2.0 (Simplified)

# Core Dependencies
livekit-agents[openai]>=1.2.15
python-dotenv>=1.0.0

# LiveKit Plugins
livekit-plugins-soniox>=0.1.0              # Soniox STT (Arabic & English)
livekit-plugins-elevenlabs>=1.2.0          # ElevenLabs TTS
livekit-plugins-silero>=1.2.0              # Silero VAD
livekit-plugins-openai>=1.2.0              # OpenAI GPT-4

# HTTP Client
aiohttp>=3.9.0                             # Async HTTP client for GraphQL

# AWS S3 for video storage
boto3>=1.34.0                              # AWS SDK for S3 uploads

# Image Processing for Vision Tool
Pillow>=10.0.0                             # PIL for image processing

# Tokenization
blingfire>=0.1.8                           # Fast sentence tokenization

# Development Dependencies (optional - install with: pip install -r requirements.txt -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
